pyproject.toml
[project]
name = "tqsdk-broker-connect"
version = "0.1.0"
description = "Separate TqSDK connection service that communicates with qpto_engine via RabbitMQ"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}
authors = [
    {name = "Internal"}
]

dependencies = [
    "tqsdk==3.8.7",
    "pyyaml==6.0",
    "SQLAlchemy==2.0.34",
    "redis==6.1.0",
    "pika==1.3.2",
    "psycopg2-binary==2.9.9",
    "pandas==2.1.4",
    "loguru==0.7.2",
]

[project.scripts]
tqsdk-broker-connect = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["tqsdk_client"]
